+++ src/audio/external.h
@@ -50,7 +50,7 @@ void play_music(u8 player, u16 seqArgs, u16 fadeTimer);
 void stop_background_music(u16 seqId);
 void fadeout_background_music(u16 seqId, u16 fadeOut);
 void drop_queued_background_music(void);
-u16 get_current_background_music(void);
+u32 get_current_background_music(void);
 void play_secondary_music(u8 seqId, u8 bgMusicVolume, u8 volume, u16 fadeTimer);
 void func_80321080(u16 fadeTimer);
 void func_803210D4(u16 fadeOutTime);
