+++ src/game/object_helpers.h
@@ -189,7 +189,7 @@ void obj_translate_xyz_random(struct Object *obj, f32 rangeLength);
 void obj_translate_xz_random(struct Object *obj, f32 rangeLength);
 void cur_obj_set_pos_via_transform(void);
 void cur_obj_spawn_particles(struct SpawnParticlesInfo *info);
-s16 cur_obj_reflect_move_angle_off_wall(void);
+s32 cur_obj_reflect_move_angle_off_wall(void);
 
 #define WAYPOINT_FLAGS_END -1
 #define WAYPOINT_FLAGS_INITIALIZED 0x8000
